Faker is now working

+++ lib/db-ci-helpers.ts
@@ -1,7 +1,11 @@
 import { RemotedDatabase } from "../server/db/model";
 
 export async function clearDb(db: RemotedDatabase): Promise<void> {
+  if (process.env.NODE_ENV === "production") {
+    throw new Error("This is not supposed to run in production");
+  }
   await db.job_tags.destroy({});
   await db.job.destroy({});
+  await db.tag.destroy({});
   await db.company.destroy({});
 }
